Make git-edit-conflicts specific to conflicts
The command roughly boiled down to `git diff --name-only | uniq`, which isn't specific to just conflicted files.  It works if run when first entering a conflicted state, but if you unstage any changes it'll cause some issues.  More to the point, if run when *not* in a conflicted state, it would just open any unstaged changes, since it's just looking for unstaged changes.  We already have a good script for listing files in a conflict state (`git-conflicts`, which runs `git ls-files -u | awk '{print $4}' | sort -u`), which is already reused in `git-mark-all-resolved`.  This mirrors that behavior to be specific to conflict state, switching the script from ruby to bash.