Add missing git-wordiness script
Add the missing git-wordiness script. This includes a few changes to
work on macOS, but should still be compatible with Linux. The reverse
flag was also removed because macOS doesn't include a tac by default.

For most purposes, it should still work (but really ought to be
rewritten in a better language).